requirements.txt
# Data processing & feature engineering
pandas==2.0.3
numpy==1.24.4
tsfresh==0.20.1
scikit-learn==1.3.0
ta==0.10.2

# Deep learning
torch==2.0.1
transformers==4.30.2

# API & serving
fastapi==0.100.0
uvicorn[standard]==0.23.1
pydantic==2.0.2
redis==4.6.0
fastapi-cache2==0.2.1
slowapi==0.1.8

# Dashboard
dash==2.12.1
dash-bootstrap-components==1.4.1
dash-extensions==1.0.4
plotly==5.15.0

# MLOps
mlflow==2.7.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
requests==2.31.0
python-dotenv==1.0.0

# Testing
pytest==7.4.0
httpx==0.24.1
